Added usage for other params and smart description sense
The smart description sense fixed an issue where typing
"create "my repo" -d -p" (essentially not providing a description
even though the tag is used) wrote "-p" (in this case) as the description.
A helper method is_tag(tag, dict) was created to see if the added
description is a tag. Thus, you cannot have tags as descriptions