Fixed critical bug in remote.py
My stupid head forget to add a sys.exit in remote.py after block of code that creates and inits a repo. As you can imagine, the bug caused the script to initialize a whole new repo in the current directory. This commit is just so we know everything is ok at this point. Also, I started the code for deleting files (which I was testing). Nearly complete