Fixed some bugs, more to go
Fixed a logic error with creating directories and changing into them. Added a couple TODOs of what needs to be developed later, it's like 1 am now. Removing files locally is EXTREMELY buggy right now, but I know what's going on it's just writing the code to fix that, which we'll hopefully see in the next commit. Ok, sleep's calling