Added colour!
Used colorama module to add some colour to the output, green when stuff is created and red when asking to destroy repos. Also added some comments to make it easier to understand. That pretty much finished output. Next is to refactor some variables and method names to make it more readable. After that I should be done and ready to issue a pull request. But on that note, I'll need to update the File structure too to make it more even. And add an extra readme. Thanks for reading this commit description. If you're seeing this, you're awesome! :)